RazorProjectItem: Introduce method that returns RazorSourceDocument
This change rationalizes the RazorProjectItem.RazorSourceDocument property that was leveraged by SourceGeneratorProjectItem by introducing a RazorProjectItem.GetSource() method that, by default, just calls RazorSourceDocument.ReadFrom(...) with itself and returns the result. Overriders can provide specialized behavior.

In this change, the following RazorProjectItem subclasses override GetSource():

- SourceGeneratorProjectItem - Provides access to a pre-computed RazorSourceDocument.
- ComponentImportProjectItem - Provides access to a RazorSourceDocument memoized in a static field.
- MvcImportProjectFeature (all three) - Provides access to a RazorSourceDocument memoized in a static field.
- AdditionalImportsProjectFeature - Provides access to a RazorSourceDocument memoized in an instance field.

This ensures that the same RazorSourceDocument is used across all default imports provided by the compiler.

Note: A handful of tests that created mocks of RazorProjectItem needed to be adjusted to create TestRazorProjectItem.