Improve RazorSourceGeneratorTests a bit
The RazorSourceGeneratorTests rely on an EventListener which can make them a challenge to debug.
During debug-time, it's for the EventListener to capture spurious events, resulting in
unexpected failures. This change adds a sanity check to ensure that the EventListener only
captures events from the expected source. In addition, the events are no longer captures in a
ConcurrentQueue, since the "queue-like" behaviors were never utilitized.