IImportProjectFeature: Change accessibility to internal
This change makes the IImportProjectFeature interface internal. In addition, the SetImportFeature extension method on RazorProjectEngineBuilder has been moved to the common test library, since it's really intended to be used by tests.

This is a public API breaking change, but there are no users of these APIs in dotnet/sdk.