Fix vs-threading configuration file to point to correct type
We had taken the time to ensure that the vs-threading analyzers could look for our custom JTF assert helpers, but it stopped working when they were moved to a different namespace. This fixes that problem, which allowed the vs-threading analyzers to find a bit of suspicious code. (The suspicious code is fully intentional and there's a comment explaining it, but it's great to see the analyzer point this out!)