Rationalize RazorProjectEngineBuilder.AddDirective(...) extension method
Change the signature of AddDirective to take a ReadOnlySpan<string> and make it internal. This API is not used by the RazorSDK, so it shouldn't be considered a breaking change.

In addition, I've removed some code duplication for features that need to get a feature or create one if it doesn't exist.