ModelDirectiveTest: Fix incorrect test assertions
Both of the design-time tests fail after being updated to use code documents from the project engine. The reason is that the test was initially buggy. Previously, the design-time tests created a RazorProjectEngine with options configured for design-time. However, the RazorCodeDocument was created without those options. So, DocumentIntermediateNode.Options.DesignTime wouldn't be set, and ModelDirective.Pass.ExecuteCore(...) would follow a runtime code path. Now that the RazorCodeDocument is correctly configured, it takes the correct design-time code path and the assertions need to be updated.

IOW, the test was wrong and now it's correct.