Make RazorCodeDocument.CodeGenerationOptions non--nullable
This change ensures that RazorCodeDocument.CodeGenerationOptions is never null. This allows the legacy IRazorCodeGenerationOptionsFeature to be completely removed. This was used as a fallback in some places in the compiler when a RazorCodeDocument didn't contain RazorCodeGenerationOptions and is no longer needed.

This also means that it should now be an invariant in the compiler that DocumentIntermediateNode.Options is the same as the code-gen options from the RazorCodeDocument that produced it.

In addition, code that used DocumentIntermediateNode.Options as a fallback when RazorCodeDocument.CodeGenerationOptions was null can now be adjusted. This affects the TryComputeNamespace(...) method. That API has tests to verify the behavior when DocumentIntermediateNode.Options differs from RazorCodeDocument.CodeGenerationOptions. This change adjusts or removes those tests.