Add RazorProjectService.AddDocumentsToMiscProjectAsync for bulk adds
When the WorkspaceRootPathWatcher starts it (currently) adds all *.razor and *.cshtml files under the root path to the RazorProjectService. It does this one file path at a time, which is pretty inefficient in large solutions, requiring lots of thread hopping for updates to the ProjectSnapshotManager. This change adds a bulk method that can be called to add an array of file paths in one update.