Avoid calling TagHelperBinding.GetBoundRules(...) when unnecessary
A common pattern is to iterate over TagHelperBinding.Descriptors, call GetBoundRules for each descriptor and iterate over the rules. Because TagHelperBinding no longer contains a Dictionary, these double-loops are now O(n^2). Fortunately, it's easy to just iterate over AllBoundRules, which is what these cases are really trying to do.