SyntaxRewriter: Remove VisitToken implementation
This method is essentially a no-op. It's purpose is to walk the leading and trailing trivia of a SyntaxToken, but the Razor parser doesn't actually produce tokens with trivia!

On my machine, this method was called 10,478 times when creating a brand new Blazor Web App in VS, but it never amounted to anything except returning its input. Now, that's all it does.