GreenNode: Make ToString() to return the same text as ToFullString()
GreenNode.ToString() has an inconsistent implementation from ToFullString(). Instead of returning the the result of building a string from its subtree, it returns what appears to be a debugging value. It turns out that only one unit test depends on this ToString() result and is easy to update. So, this change makes ToString() return the same result as ToFullString(). The old ToString() implementation is used for debugger display.