Fix CSharpTestLspServer
The RoslynLanguageServer in the latest Microsoft.CodeAnalysis.LanguageServer.Protocol.dll doesn't exit gracefully when the JsonRpc it's listening to is forcibly disconnected. Unfortunately, that's exactly how our CSharpTestLspServer shuts down the RoslynLanguageServer that it creates. Ultimately, the Razor EA should be updated to provide a IAsyncDisposable implementation that can be called to gracefully shut down the language server. However, in the meantime, this change uses reflection to get at the underlying RoslynLanguageServer and call ShutdownAsync and ExitAsync on it before disposing the server-side JsonRpc.