Reduce allocations in TagHelperParseTreeRewriter.Rewrite
Customer profile is showing a very large number of allocations in the GetAllDiagnostics call, due to the yield state machinery. Add internal method to TagHelperDescriptor that allows callers to pass in a PooledArrayBuilder for the diagnostics to be added directly to, instead of using "return yield".

Partial fix for https://github.com/dotnet/roslyn/issues/78639