Reduce costs in ComponentDirectiveVisitor.VisitRazorDirective (#11881)
* Reduce costs in ComponentDirectiveVisitor.VisitRazorDirective

Customer feedback ticket shows ~40% of CPU costs are in this method. Of the 32 seconds of CPU time in this method, 21 seconds are in it's calls to GetTypeNamespace and GetSpanWithoutGlobalPrefix.

VisitRazorDirective is called many times (we'll call that number 'l'), depending on the number of directives in the page. Each call costs O(m x n) where m is the number of import statements in the page and n is the number of directives that aren't fully qualified. Evidently, for this customer on this page, this O(l * m * n) is too much.

1) GetTypeNamespace -- Reduced to O(n) calls by storing it's result in _nonFullyQualifiedComponents
2) GetSpanWithoutGlobalPrefix -- Reduced to O(l * m) by moving outside the innnermost loop
3) Reduced innermost loop to only walk taghelpers in the same namespace

Fixes: https://developercommunity.visualstudio.com/t/Razor-Development:-30-Second-CPU-Churn-A/10904811