Fix #11712
A bug introduced in ac6e32a73fb93b3c699269fa432efac38fe4d5fe causes RazorSourceDocuments for imports to be created with the wrong file paths. Instead of a pointing to a physical file path on disk, they contain a relative path, which causes incorrect code generation. To fix this, call the GetSourceAsync(...) helper for the import. That will populate the RazorSourceDocument with the correct paths.