EnumerableExtensions.TryGetCount: Test IReadOnlyCollection<T> on .NET
It turns out that Enumerable.TryGetNonEnumeratedCount(...) only tests for ICollection and ICollection<T>. This means that it returns false for IReadOnlyList<T> and IReadOnlyCollection<T> implementations. So, we need to test for IReadOnlyCollection<T> specifically.

Note: There is effort in .NET 10 to fix this, but it isn't in yet. Once it is, we can update this again.