Reduce allocations during writing of the filepath in the CodeWriter
Previously, the code would do a string.Replace on the filePath to allocate a new windows-friendly path. Instead, we can just separate the path into Spans and add those to the CodeWriter (separated by backslashes)

During the typing scenario in the razor editing speedometer test, the string.Replace accounts for about 0.4% of allocations.