Create a specialized pool for arrays of ClassifiedSpanInternal
These arrays show up as a fairly large allocation in the typing scenario of the razor speedometer test. There are 150 MB (2.1%)of allocations of ClassifiedSpanInternal[] in the profile, but only 2.1 MB of ImmutableArray<ClassifiedSpanInternal>. This should be partly addressed by the earlier switch from DrainToImmutable use ToImmutableAndClear. However, this is also indicative that the 512 entry threshold for ArrayBuilderPool isn't sufficient for ClassifiedSpanInternal array counts. Roslyn has experienced this too and utilizes separate array pools for classification/tagging performance. The new pool has a threshold of 16K, much larger than the 512 entry limit. I tested this against a 60 KB razor file and was hitting about 7K entries, so this seemed like a pretty reasonable value.