ChildSyntaxList.ItemInternal optimization
This is a port of https://github.com/dotnet/roslyn/pull/73650 to Razor

The general idea behind this PR is that calls to getting items in a syntax list are typically done sequentially, and thus walking the full children of a node is often not necessary.

See the Roslyn PR for detailed performance analysis that was done upon the roslyn investigation. Noted that