Optimize allocations under RuntimeNodeWriter.WriteHtmlContent
This method was creating a StringBuilder with unknown size allocating both a final string and char array along with resize allocations. Instead, calculate the result length and calculate a single char array. This reduced allocations in the typing scenario in the razor lsp speedometer test under this method from about 1.1% to 0.4%.