Fix nasty bug in CodeWriter's TextReader implementation
`CodeWriter.Reader.Read(char[] buffer, int index, int count)` is implemented incorrectly when called after called after it has no more data to return. In that case, it returns -1 rather than 0. Because this method is called in a loop by `TextReader.ReadBlock(...)` while the result is greater than 0, -1 does end the loop. However, It causes -1 to be *added* to the final length returned by `ReadBlock`, reducing it by 1. This causes the `LargeText` returned to have a length that is 1 fewer than expected.

In addition, the `TextReader` implementaion failed to reduce `_remainingLength` as expected. This is now fixed as well.