Improves IntermediateNodeWalker efficiency
Refactor `IntermediateNodeWalker` to use a simple array+pointer as a stack for tracking ancestors. The stack grows downward so that the ancestors are in the correct order expected to be returned from the `Ancestors` property. Previously, this was achieved using a `List<IntermediateNode>` and inserting/removing at the start of the list for each ancestor, which causes a lot of array copying.