Change Build(Enhanced)LinePragma extensions to target CodeRenderingContext
Because `CodeRenderingContext` owns the `CodeWriter` instance used for code gen, and the `BuildLinePragma` and `BuildEnhancedLinePragma` extension methods target `CodeWriter` and also take a `CodeRenderingContext`, we can change them can target `CodeRenderingContext`.

In addition, this change changes the type that's returned to readonly ref struct. This requires a bit of adjustment to some of the calling code.