CR Feedback: Revert behavior change
When refactoring DesignTimeNodeWriter.WriteCSharpCodeAttributeValue(...), I assumed that the compiler wouldn't ever produce CSharpIntermediateTokens with non-whitespace content but no source mapping. However, this is technically a behavior change and we should write such tokens without a source mapping as a single line.

I've added a Debug.Fail(...) to catch this scenario at debug-time as it seems like an invariant that non-whitespace tokens *should* have a valid Source.