Don't use dictionary and string.Replace to build decoded text
TryDecodeHtmlEntities uses a pretty inefficient for building decoded text after HTML character entities are processed. It scans through the content and tries to decode character entities when it encounters a '&' character. It adds each character entity and its replacement to a dictionary. Then, when it's done scanning, it looks through the dictionary and calls string.Replace for each one.

This results in a string allocation per different character entity found. So, if the content contains both '&gt;' and '&lt;'. There will be two string allocations to produce the decoded text.

The change refactors TryDecodeHtmlEntities to avoid these string allocations. Instead of using a dictionary, it uses a MemoryBuilder<ReadOnlyMemory<char>> to track the chunks needed to build the final decoded chunks.