Improve TagHelperObjectBuilderCollection.ToImmutable perf in a specific case
Locally, I see about 2/3 of the non-empty ToImmutable calls be of size 1. This just handles that case directly without the need for the pooled data structures or adding to a hashset.

The ToImmutable call shows up as 2.7% of CPU samples during the html completion scenario in the razor cohosting completion speedometer test.

Going to get a perf run to make sure this improves things (ever so slightly)