Don't use ImmutableHashSet and ImmutableDictionary in CSharpCodeParser
These two collection types have O(log N) look-up instead of O(1) and are best used in scenarios that take advantage of them as persistent data structures. However, that's not really the case here, since they're built up in the constructor and never modified during the lifetime of CSharpCodeParser. This change removes the immutable data structions in favor of Dictionary<TKey, TValue> and introduces a simple KeywordSet abstraction that wraps a FrozenSet<string> or HashSet<string>.