Ensure RazorVSInternalCompletionParams is used for serialization of completion requests (#12271)
* Ensure RazorVSInternalCompletionParams is used for serialization of completion requests

Without this type being used, serialization loses the VSInternalCompletionContext.InvokeKind, which is critical for webtools to determine the correct range to replace.

Without this, webtools indicates a range to be replaced based on the assumption that this is a forced completion session, which in turns affects how vseditor determines filtering of items in the completion list.

Fixes https://dev.azure.com/devdiv/DevDiv/_workitems/edit/2581568

* 1) Fix ci build errors
2) Propagate RazorVSInternalCompletionParams usage a bit further out per David's suggestion