Add GreenNode.GetEnumerator() to perform depth-first traversal
Add GreenNode.GetEnumerator() method that returns a ref struct-based Enumerator to perform a depth-first traversal using the target GreenNode as the root. SyntaxList nodes are transparent and not returned from the enumerator.

Unit tests are included along with some helpers to make constructing GreenNodes a bit easier. The helpers include:

- A handful of SyntaxFactory methods to avoid passing extra arguments.
- A [CollectionBuilder] attribute for SyntaxList<TNode>
- SyntaxList.Create(...) methods for the [CollectionBuilder] to target.