Rework GreenNode.ToString() for efficiency
This change removes the StringWriter machinery from GreenNode. It only adds overhead. In its place is a more efficient implementation using string.Create. This avoids copying the characters into a StringWriter and then allocating a new string from the writer, which copies all of the characters *again* into a new string.

Because we know the width of a GreenNode, we know the length of the string it will produce.