Remove SyntaxTokenCache.Entry (#12370)
While debugging some related code, I noticed this class didn't use a lock when getting the Entry, which smelled to me like a possible torn read/writes issue (and those suck). I mentioned this to Dustin to make sure I wasn't being goofy, and he thought that might be a concern here too, but then raised the valid point that the reason for it being a struct (to have a hash in addition to the SyntaxToken) wasn't even necessary.