Reduce allocated strings in CodeWriter (#12366)
For nearly all cases, the result of GetIndentString calls can be returned without allocation, but slicing into a string of spaces or tabs. ComputeIndent accounts for 11.0 MB (0.4%) of total allocations and 27 ms in in the ScrollingAndTypingInCohosting speedometer test. With this change, those numbers go to 0.0 MB (0.0%) and 13 ms respectively.

Test insertion: https://dev.azure.com/devdiv/DevDiv/_git/VS/pullrequest/680751