Introduce Optional<T>
This is based on Roslyn's Optional<T>, which is similar to Nullable<T> except that it can hold reference types and does not throw when `Value` is accessed and `HasValue` is false. It is particular useful for long lists of optional parameters.

Note: This type is public because Razor was currently using Roslyn's Optional<T> in a couple of public places, and those are now using Razor's Optional<T>.