Better handle parameter attribute directive completion (#12473)
Modifies completion in several ways:

Makes ':' no longer a commit character. This being a commit character was causing difficulties as it and '=' require fundamentally different completion commit behavior, and LSP doesn't support that for a single completion item.
Makes parameter completions show up in attribute name contexts too. This is nice for discoverability.

Fixes: https://devdiv.visualstudio.com/DevDiv/_workitems/edit/2607948