Add cache to TagHelperDocumentContext
TagHelperBinders are expensive to create, and Razor often creates them for the same set of tag helpers. This change introduces a cache to avoid creating a new TagHelperDocumentContext (and therefore, a new TagHelperBinder) for the same TagHelperCollection.

Note: This requires adding some deduping logic to Rename, which started creating duplicate ranges with this change.